@@ -4,7 +4,7 @@ requirements = """pandas
 ipython
 numpy
 opencv-python
-matplotlib
+matplotlib>=3.5
 h5py<3.0.0
 scikit-image
 tensorflow<=1.15
@@ -17,13 +17,13 @@ keras==2.3
 setuptools.setup(
     name='granny',
     packages=setuptools.find_packages(),
-    url="",
+    url="https://github.com/SystemsGenetics/granny",
     version='0.0.1',
     description='GRANNY is an implementation of Mask-RCNN and image processing techniques,\
         developed by the Ficklin Research Program, to rate disorder severity in Granny Smith apple.',
     author='Nhan H Nguyen',
     license='GNU General Public License v3.0',
-    python_requires='==3.7',
+    python_requires='<3.8',
     install_requires=requirements,
     entry_points={'console_scripts': [
         'granny = GRANNY.command:main',
